move rle encode decode functions to dataset/utils.py
